refactor: moderate reorganizing
