+++ cmd/loadtest/app.go
@@ -2,26 +2,23 @@ package loadtest
 
 import (
 	"crypto/ecdsa"
-	"github.com/maticnetwork/polygon-cli/rpctypes"
-	"github.com/spf13/cobra"
-	"net/url"
-	"regexp"
-)
-
-import (
+	_ "embed"
 	"fmt"
 	"math/big"
+	"math/rand"
+	"net/url"
+	"regexp"
 	"strings"
+	"sync"
 	"time"
-
-	_ "embed"
 
 	gssignature "github.com/centrifuge/go-substrate-rpc-client/v4/signature"
 	gstypes "github.com/centrifuge/go-substrate-rpc-client/v4/types"
-
 	ethcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/maticnetwork/polygon-cli/rpctypes"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
+	"github.com/spf13/cobra"
 )
 
 type (
@@ -97,6 +94,76 @@ type (
 		Pending any `json:"pending"`
 		Queued  any `json:"queued"`
 	}
+)
+
+var (
+	//go:embed usage.md
+	usage               string
+	inputLoadTestParams loadTestParams
+	loadTestResults     []loadTestSample
+	loadTestResutsMutex sync.RWMutex
+	validLoadTestModes  = []string{
+		loadTestModeTransaction,
+		loadTestModeDeploy,
+		loadTestModeCall,
+		loadTestModeFunction,
+		loadTestModeInc,
+		loadTestModeStore,
+		loadTestModeERC20,
+		loadTestModeERC721,
+		loadTestModePrecompiledContracts,
+		loadTestModePrecompiledContract,
+		// r should be last to exclude it from random mode selection
+		loadTestModeRandom,
+	}
+
+	hexwords = []byte{
+		0x00, 0x0F, 0xF1, 0xCE,
+		0x00, 0xBA, 0xB1, 0x0C,
+		0x1B, 0xAD, 0xB0, 0x02,
+		0x8B, 0xAD, 0xF0, 0x0D,
+		0xAB, 0xAD, 0xBA, 0xBE,
+		0xB1, 0x05, 0xF0, 0x0D,
+		0xB1, 0x6B, 0x00, 0xB5,
+		0x0B, 0x00, 0xB1, 0x35,
+		0xBA, 0xAA, 0xAA, 0xAD,
+		0xBA, 0xAD, 0xF0, 0x0D,
+		0xBA, 0xD2, 0x22, 0x22,
+		0xBA, 0xDD, 0xCA, 0xFE,
+		0xCA, 0xFE, 0xB0, 0xBA,
+		0xB0, 0xBA, 0xBA, 0xBE,
+		0xBE, 0xEF, 0xBA, 0xBE,
+		0xC0, 0x00, 0x10, 0xFF,
+		0xCA, 0xFE, 0xBA, 0xBE,
+		0xCA, 0xFE, 0xD0, 0x0D,
+		0xCE, 0xFA, 0xED, 0xFE,
+		0x0D, 0x15, 0xEA, 0x5E,
+		0xDA, 0xBB, 0xAD, 0x00,
+		0xDE, 0xAD, 0x2B, 0xAD,
+		0xDE, 0xAD, 0xBA, 0xAD,
+		0xDE, 0xAD, 0xBA, 0xBE,
+		0xDE, 0xAD, 0xBE, 0xAF,
+		0xDE, 0xAD, 0xBE, 0xEF,
+		0xDE, 0xAD, 0xC0, 0xDE,
+		0xDE, 0xAD, 0xDE, 0xAD,
+		0xDE, 0xAD, 0xD0, 0x0D,
+		0xDE, 0xAD, 0xFA, 0x11,
+		0xDE, 0xAD, 0x10, 0xCC,
+		0xDE, 0xAD, 0xFE, 0xED,
+		0xDE, 0xCA, 0xFB, 0xAD,
+		0xDE, 0xFE, 0xC8, 0xED,
+		0xD0, 0xD0, 0xCA, 0xCA,
+		0xE0, 0x11, 0xCF, 0xD0,
+		0xFA, 0xCE, 0xFE, 0xED,
+		0xFB, 0xAD, 0xBE, 0xEF,
+		0xFE, 0xE1, 0xDE, 0xAD,
+		0xFE, 0xED, 0xBA, 0xBE,
+		0xFE, 0xED, 0xC0, 0xDE,
+		0xFF, 0xBA, 0xDD, 0x11,
+		0xF0, 0x0D, 0xBA, 0xBE,
+	}
+
+	randSrc *rand.Rand
 )
 
 // LoadtestCmd represents the loadtest command
