+++ cmd/loadtest/loadtest.go
@@ -3,34 +3,21 @@ package loadtest
 import (
 	"context"
 	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"math"
 	"math/big"
 	"math/rand"
 	"os"
 	"os/signal"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/maticnetwork/polygon-cli/metrics"
-	"github.com/maticnetwork/polygon-cli/rpctypes"
-	"github.com/maticnetwork/polygon-cli/util"
-	"golang.org/x/exp/constraints"
-	"golang.org/x/text/language"
-	"golang.org/x/text/message"
-	"golang.org/x/text/number"
-
 	_ "embed"
-
 	gsrpc "github.com/centrifuge/go-substrate-rpc-client/v4"
-	gssignature "github.com/centrifuge/go-substrate-rpc-client/v4/signature"
-	gstypes "github.com/centrifuge/go-substrate-rpc-client/v4/types"
+	"github.com/maticnetwork/polygon-cli/metrics"
 
 	ethereum "github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
@@ -41,7 +28,6 @@ import (
 	ethrpc "github.com/ethereum/go-ethereum/rpc"
 
 	"github.com/maticnetwork/polygon-cli/contracts"
-	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"golang.org/x/time/rate"
 )
@@ -61,76 +47,6 @@ const (
 
 	codeQualitySeed       = "code code code code code code code code code code code quality"
 	codeQualityPrivateKey = "42b6e34dc21598a807dc19d7784c71b2a7a01f6480dc6f58258f78e539f1a1fa"
-)
-
-var (
-	//go:embed usage.md
-	usage               string
-	inputLoadTestParams loadTestParams
-	loadTestResults     []loadTestSample
-	loadTestResutsMutex sync.RWMutex
-	validLoadTestModes  = []string{
-		loadTestModeTransaction,
-		loadTestModeDeploy,
-		loadTestModeCall,
-		loadTestModeFunction,
-		loadTestModeInc,
-		loadTestModeStore,
-		loadTestModeERC20,
-		loadTestModeERC721,
-		loadTestModePrecompiledContracts,
-		loadTestModePrecompiledContract,
-		// r should be last to exclude it from random mode selection
-		loadTestModeRandom,
-	}
-
-	hexwords = []byte{
-		0x00, 0x0F, 0xF1, 0xCE,
-		0x00, 0xBA, 0xB1, 0x0C,
-		0x1B, 0xAD, 0xB0, 0x02,
-		0x8B, 0xAD, 0xF0, 0x0D,
-		0xAB, 0xAD, 0xBA, 0xBE,
-		0xB1, 0x05, 0xF0, 0x0D,
-		0xB1, 0x6B, 0x00, 0xB5,
-		0x0B, 0x00, 0xB1, 0x35,
-		0xBA, 0xAA, 0xAA, 0xAD,
-		0xBA, 0xAD, 0xF0, 0x0D,
-		0xBA, 0xD2, 0x22, 0x22,
-		0xBA, 0xDD, 0xCA, 0xFE,
-		0xCA, 0xFE, 0xB0, 0xBA,
-		0xB0, 0xBA, 0xBA, 0xBE,
-		0xBE, 0xEF, 0xBA, 0xBE,
-		0xC0, 0x00, 0x10, 0xFF,
-		0xCA, 0xFE, 0xBA, 0xBE,
-		0xCA, 0xFE, 0xD0, 0x0D,
-		0xCE, 0xFA, 0xED, 0xFE,
-		0x0D, 0x15, 0xEA, 0x5E,
-		0xDA, 0xBB, 0xAD, 0x00,
-		0xDE, 0xAD, 0x2B, 0xAD,
-		0xDE, 0xAD, 0xBA, 0xAD,
-		0xDE, 0xAD, 0xBA, 0xBE,
-		0xDE, 0xAD, 0xBE, 0xAF,
-		0xDE, 0xAD, 0xBE, 0xEF,
-		0xDE, 0xAD, 0xC0, 0xDE,
-		0xDE, 0xAD, 0xDE, 0xAD,
-		0xDE, 0xAD, 0xD0, 0x0D,
-		0xDE, 0xAD, 0xFA, 0x11,
-		0xDE, 0xAD, 0x10, 0xCC,
-		0xDE, 0xAD, 0xFE, 0xED,
-		0xDE, 0xCA, 0xFB, 0xAD,
-		0xDE, 0xFE, 0xC8, 0xED,
-		0xD0, 0xD0, 0xCA, 0xCA,
-		0xE0, 0x11, 0xCF, 0xD0,
-		0xFA, 0xCE, 0xFE, 0xED,
-		0xFB, 0xAD, 0xBE, 0xEF,
-		0xFE, 0xE1, 0xDE, 0xAD,
-		0xFE, 0xED, 0xBA, 0xBE,
-		0xFE, 0xED, 0xC0, 0xDE,
-		0xFF, 0xBA, 0xDD, 0x11,
-		0xF0, 0x0D, 0xBA, 0xBE,
-	}
-
-	randSrc *rand.Rand
 )
 
 func initializeLoadTestParams(ctx context.Context, c *ethclient.Client) error {
@@ -331,35 +247,6 @@ func runLoadTest(ctx context.Context) error {
 	}
 	log.Info().Msg("Finished")
 	return nil
-}
-
-func printResults(lts []loadTestSample) {
-	if len(lts) == 0 {
-		log.Error().Msg("No results recorded")
-		return
-	}
-
-	log.Info().Msg("* Results")
-	log.Info().Int("samples", len(lts)).Msg("Samples")
-
-	var startTime = lts[0].RequestTime
-	var endTime = lts[len(lts)-1].RequestTime
-	var meanWait float64
-	var totalWait float64 = 0
-	var numErrors uint64 = 0
-
-	for _, s := range lts {
-		if s.IsError {
-			numErrors += 1
-		}
-		totalWait = float64(s.WaitTime.Seconds()) + totalWait
-	}
-	meanWait = totalWait / float64(len(lts))
-
-	log.Info().Time("startTime", startTime).Msg("Start")
-	log.Info().Time("endTime", endTime).Msg("End")
-	log.Info().Float64("meanWait", meanWait).Msg("Mean Wait")
-	log.Info().Uint64("numErrors", numErrors).Msg("Num errors")
 }
 
 func convHexToUint64(hexString string) (uint64, error) {
@@ -691,33 +578,6 @@ func mainLoop(ctx context.Context, c *ethclient.Client, rpc *ethrpc.Client) erro
 		}
 	}
 	return nil
-}
-
-func lightSummary(ctx context.Context, c *ethclient.Client, rpc *ethrpc.Client, startBlockNumber, startNonce, endBlockNumber, endNonce uint64, rl *rate.Limiter) {
-	startBlock, err := c.BlockByNumber(ctx, new(big.Int).SetUint64(startBlockNumber))
-	if err != nil {
-		log.Error().Err(err).Msg("unable to get start block for light summary")
-		return
-	}
-	endBlock, err := c.BlockByNumber(ctx, new(big.Int).SetUint64(endBlockNumber))
-	if err != nil {
-		log.Error().Err(err).Msg("unable to get end block for light summary")
-		return
-	}
-	endTime := time.Unix(int64(endBlock.Time()), 0)
-	startTime := time.Unix(int64(startBlock.Time()), 0)
-
-	testDuration := endTime.Sub(startTime)
-	tps := float64(len(loadTestResults)) / testDuration.Seconds()
-
-	log.Info().
-		Time("firstBlockTime", startTime).
-		Time("lastBlockTime", endTime).
-		Int("transactionCount", len(loadTestResults)).
-		Float64("testDuration", testDuration.Seconds()).
-		Float64("tps", tps).
-		Float64("final rate limit", float64(rl.Limit())).
-		Msg("rough test summary (ignores errors)")
 }
 
 func blockUntilSuccessful(ctx context.Context, c *ethclient.Client, f func() error, numberOfBlocksToWaitFor, blockInterval uint64) error {
@@ -1081,6 +941,12 @@ func loadtestERC721(ctx context.Context, c *ethclient.Client, nonce uint64, erc7
 
 	return
 }
+func loadtestNotImplemented(ctx context.Context, c *gsrpc.SubstrateAPI, nonce uint64) (t1 time.Time, t2 time.Time, err error) {
+	t1 = time.Now()
+	t2 = time.Now()
+	err = fmt.Errorf("this method is not implemented")
+	return
+}
 
 func recordSample(goRoutineID, requestID int64, err error, start, end time.Time, nonce uint64) {
 	s := loadTestSample{}
@@ -1122,256 +988,6 @@ func getRandomAddress() *ethcommon.Address {
 	}
 	realAddr := ethcommon.BytesToAddress(addr)
 	return &realAddr
-}
-
-func availLoop(ctx context.Context, c *gsrpc.SubstrateAPI) error {
-	var err error
-
-	ltp := inputLoadTestParams
-	log.Trace().Interface("Input Params", ltp).Msg("Params")
-
-	routines := *ltp.Concurrency
-	requests := *ltp.Requests
-	currentNonce := uint64(0) // *ltp.CurrentNonce
-	chainID := new(big.Int).SetUint64(*ltp.ChainID)
-	privateKey := ltp.ECDSAPrivateKey
-	mode := *ltp.Mode
-
-	_ = chainID
-	_ = privateKey
-
-	meta, err := c.RPC.State.GetMetadataLatest()
-	if err != nil {
-		return err
-	}
-
-	genesisHash, err := c.RPC.Chain.GetBlockHash(0)
-	if err != nil {
-		return err
-	}
-
-	key, err := gstypes.CreateStorageKey(meta, "System", "Account", ltp.FromAvailAddress.PublicKey, nil)
-	if err != nil {
-		log.Error().Err(err).Msg("Could not create storage key")
-		return err
-	}
-
-	var accountInfo gstypes.AccountInfo
-	ok, err := c.RPC.State.GetStorageLatest(key, &accountInfo)
-	if err != nil {
-		log.Error().Err(err).Msg("Could not load storage")
-		return err
-	}
-	if !ok {
-		err = fmt.Errorf("loaded storage is not okay")
-		log.Error().Err(err).Msg("Loaded storage is not okay")
-		return err
-	}
-
-	currentNonce = uint64(accountInfo.Nonce)
-
-	rl := rate.NewLimiter(rate.Limit(*ltp.RateLimit), 1)
-	if *ltp.RateLimit <= 0.0 {
-		rl = nil
-	}
-
-	var currentNonceMutex sync.Mutex
-
-	var i int64
-
-	var wg sync.WaitGroup
-	for i = 0; i < routines; i = i + 1 {
-		log.Trace().Int64("routine", i).Msg("Starting Thread")
-		wg.Add(1)
-		go func(i int64) {
-			var j int64
-			var startReq time.Time
-			var endReq time.Time
-
-			for j = 0; j < requests; j = j + 1 {
-
-				if rl != nil {
-					err = rl.Wait(ctx)
-					if err != nil {
-						log.Error().Err(err).Msg("Encountered a rate limiting error")
-					}
-				}
-				currentNonceMutex.Lock()
-				myNonceValue := currentNonce
-				currentNonce = currentNonce + 1
-				currentNonceMutex.Unlock()
-
-				localMode := mode
-				// if there are multiple modes, iterate through them, 'r' mode is supported here
-				if len(mode) > 1 {
-					localMode = string(mode[int(i+j)%(len(mode))])
-				}
-				// if we're doing random, we'll just pick one based on the current index
-				if localMode == loadTestModeRandom {
-					localMode = validLoadTestModes[int(i+j)%(len(validLoadTestModes)-1)]
-				}
-				// this function should probably be abstracted
-				switch localMode {
-				case loadTestModeTransaction:
-					startReq, endReq, err = loadtestAvailTransfer(ctx, c, myNonceValue, meta, genesisHash)
-				case loadTestModeDeploy:
-					startReq, endReq, err = loadtestNotImplemented(ctx, c, myNonceValue)
-				case loadTestModeCall:
-					startReq, endReq, err = loadtestNotImplemented(ctx, c, myNonceValue)
-				case loadTestModeFunction:
-					startReq, endReq, err = loadtestNotImplemented(ctx, c, myNonceValue)
-				case loadTestModeInc:
-					startReq, endReq, err = loadtestNotImplemented(ctx, c, myNonceValue)
-				case loadTestModeStore:
-					startReq, endReq, err = loadtestAvailStore(ctx, c, myNonceValue, meta, genesisHash)
-				default:
-					log.Error().Str("mode", mode).Msg("We've arrived at a load test mode that we don't recognize")
-				}
-				recordSample(i, j, err, startReq, endReq, myNonceValue)
-				if err != nil {
-					log.Trace().Err(err).Msg("Recorded an error while sending transactions")
-				}
-
-				log.Trace().Int64("routine", i).Str("mode", localMode).Int64("request", j).Msg("Request")
-			}
-			wg.Done()
-		}(i)
-
-	}
-	log.Trace().Msg("Finished starting go routines. Waiting..")
-	wg.Wait()
-	return nil
-
-}
-
-func loadtestNotImplemented(ctx context.Context, c *gsrpc.SubstrateAPI, nonce uint64) (t1 time.Time, t2 time.Time, err error) {
-	t1 = time.Now()
-	t2 = time.Now()
-	err = fmt.Errorf("this method is not implemented")
-	return
-}
-
-func initAvailTestParams(ctx context.Context, c *gsrpc.SubstrateAPI) error {
-	toAddr, err := gstypes.NewMultiAddressFromHexAccountID(*inputLoadTestParams.ToAddress)
-	if err != nil {
-		log.Error().Err(err).Msg("Unable to create new multi address")
-		return err
-	}
-
-	if *inputLoadTestParams.PrivateKey == codeQualityPrivateKey {
-		// Avail keys can use the same seed but the way the key is derived is different
-		*inputLoadTestParams.PrivateKey = codeQualitySeed
-	}
-
-	kp, err := gssignature.KeyringPairFromSecret(*inputLoadTestParams.PrivateKey, uint8(*inputLoadTestParams.ChainID))
-	if err != nil {
-		log.Error().Err(err).Msg("Could not create key pair")
-		return err
-	}
-
-	amt, err := hexToBigInt(*inputLoadTestParams.HexSendAmount)
-	if err != nil {
-		log.Error().Err(err).Msg("Couldn't parse send amount")
-		return err
-	}
-
-	rv, err := c.RPC.State.GetRuntimeVersionLatest()
-	if err != nil {
-		log.Error().Err(err).Msg("Couldn't get runtime version")
-		return err
-	}
-
-	inputLoadTestParams.AvailRuntime = rv
-	inputLoadTestParams.SendAmount = amt
-	inputLoadTestParams.FromAvailAddress = &kp
-	inputLoadTestParams.ToAvailAddress = &toAddr
-	return nil
-}
-
-func loadtestAvailTransfer(ctx context.Context, c *gsrpc.SubstrateAPI, nonce uint64, meta *gstypes.Metadata, genesisHash gstypes.Hash) (t1 time.Time, t2 time.Time, err error) {
-	ltp := inputLoadTestParams
-
-	toAddr := *ltp.ToAvailAddress
-	if *ltp.ToRandom {
-		pk := make([]byte, 32)
-		_, err = randSrc.Read(pk)
-		if err != nil {
-			// For some reason weren't able to read the random data
-			log.Error().Msg("Sending to random is not implemented for substrate yet")
-		} else {
-			toAddr = gstypes.NewMultiAddressFromAccountID(pk)
-		}
-
-	}
-
-	gsCall, err := gstypes.NewCall(meta, "Balances.transfer", toAddr, gstypes.NewUCompact(ltp.SendAmount))
-	if err != nil {
-		return
-	}
-
-	ext := gstypes.NewExtrinsic(gsCall)
-	rv := ltp.AvailRuntime
-	kp := *inputLoadTestParams.FromAvailAddress
-
-	o := gstypes.SignatureOptions{
-		BlockHash:          genesisHash,
-		Era:                gstypes.ExtrinsicEra{IsMortalEra: false, IsImmortalEra: true},
-		GenesisHash:        genesisHash,
-		Nonce:              gstypes.NewUCompactFromUInt(uint64(nonce)),
-		SpecVersion:        rv.SpecVersion,
-		Tip:                gstypes.NewUCompactFromUInt(0),
-		TransactionVersion: rv.TransactionVersion,
-	}
-
-	err = ext.Sign(kp, o)
-	if err != nil {
-		return
-	}
-
-	t1 = time.Now()
-	defer func() { t2 = time.Now() }()
-	_, err = c.RPC.Author.SubmitExtrinsic(ext)
-	return
-}
-
-func loadtestAvailStore(ctx context.Context, c *gsrpc.SubstrateAPI, nonce uint64, meta *gstypes.Metadata, genesisHash gstypes.Hash) (t1 time.Time, t2 time.Time, err error) {
-	ltp := inputLoadTestParams
-
-	inputData := make([]byte, *ltp.ByteCount)
-	_, _ = hexwordRead(inputData)
-
-	gsCall, err := gstypes.NewCall(meta, "DataAvailability.submit_data", gstypes.NewBytes([]byte(inputData)))
-	if err != nil {
-		return
-	}
-
-	// Create the extrinsic
-	ext := gstypes.NewExtrinsic(gsCall)
-
-	rv := ltp.AvailRuntime
-
-	kp := *inputLoadTestParams.FromAvailAddress
-
-	o := gstypes.SignatureOptions{
-		BlockHash:          genesisHash,
-		Era:                gstypes.ExtrinsicEra{IsMortalEra: false, IsImmortalEra: true},
-		GenesisHash:        genesisHash,
-		Nonce:              gstypes.NewUCompactFromUInt(uint64(nonce)),
-		SpecVersion:        rv.SpecVersion,
-		Tip:                gstypes.NewUCompactFromUInt(100),
-		TransactionVersion: rv.TransactionVersion,
-	}
-	// Sign the transaction using Alice's default account
-	err = ext.Sign(kp, o)
-	if err != nil {
-		return
-	}
-
-	// Send the extrinsic
-	t1 = time.Now()
-	defer func() { t2 = time.Now() }()
-	_, err = c.RPC.Author.SubmitExtrinsic(ext)
-	return
 }
 
 func configureTransactOpts(tops *bind.TransactOpts) *bind.TransactOpts {
@@ -1433,386 +1049,6 @@ func waitForFinalBlock(ctx context.Context, c *ethclient.Client, rpc *ethrpc.Cli
 
 	log.Trace().Uint64("currentNonce", currentNonce).Uint64("startblock", startBlockNumber).Uint64("endblock", lastBlockNumber).Msg("It looks like all transactions have been mined")
 	return lastBlockNumber, nil
-}
-
-func summarizeTransactions(ctx context.Context, c *ethclient.Client, rpc *ethrpc.Client, startBlockNumber, startNonce, lastBlockNumber, endNonce uint64) error {
-	ltp := inputLoadTestParams
-	var err error
-
-	log.Trace().Msg("Starting block range capture")
-	// confirm start block number is ok
-	_, err = c.BlockByNumber(ctx, new(big.Int).SetUint64(startBlockNumber))
-	if err != nil {
-		return err
-	}
-	rawBlocks, err := util.GetBlockRange(ctx, startBlockNumber, lastBlockNumber, rpc)
-	if err != nil {
-		return err
-	}
-	// TODO: Add some kind of decimation to avoid summarizing for 10 minutes?
-	batchSize := *ltp.BatchSize
-	goRoutineLimit := *ltp.Concurrency
-	var txGroup sync.WaitGroup
-	threadPool := make(chan bool, goRoutineLimit)
-	log.Trace().Msg("Starting tx receipt capture")
-	rawTxReceipts := make([]*json.RawMessage, 0)
-	var rawTxReceiptsLock sync.Mutex
-	var txGroupErr error
-
-	startReceipt := time.Now()
-	for k := range rawBlocks {
-		threadPool <- true
-		txGroup.Add(1)
-		go func(b *json.RawMessage) {
-			var receipt []*json.RawMessage
-			receipt, err = util.GetReceipts(ctx, []*json.RawMessage{b}, rpc, batchSize)
-			if err != nil {
-				txGroupErr = err
-				return
-			}
-			rawTxReceiptsLock.Lock()
-			rawTxReceipts = append(rawTxReceipts, receipt...)
-			rawTxReceiptsLock.Unlock()
-			<-threadPool
-			txGroup.Done()
-		}(rawBlocks[k])
-	}
-
-	endReceipt := time.Now()
-	txGroup.Wait()
-	if txGroupErr != nil {
-		log.Error().Err(err).Msg("One of the threads fetching tx receipts failed")
-		return err
-	}
-
-	blocks := make([]rpctypes.RawBlockResponse, 0)
-	for _, b := range rawBlocks {
-		var block rpctypes.RawBlockResponse
-		err = json.Unmarshal(*b, &block)
-		if err != nil {
-			log.Error().Err(err).Msg("Error decoding block response")
-			return err
-		}
-		blocks = append(blocks, block)
-	}
-	log.Info().Int("len", len(blocks)).Msg("Block summary")
-
-	txReceipts := make([]rpctypes.RawTxReceipt, 0)
-	log.Trace().Int("len", len(rawTxReceipts)).Msg("Raw receipts")
-	for _, r := range rawTxReceipts {
-		if isEmptyJSONResponse(r) {
-			continue
-		}
-		var receipt rpctypes.RawTxReceipt
-		err = json.Unmarshal(*r, &receipt)
-		if err != nil {
-			log.Error().Err(err).Msg("Error decoding tx receipt response")
-			return err
-		}
-		txReceipts = append(txReceipts, receipt)
-	}
-	log.Info().Int("len", len(txReceipts)).Msg("Receipt summary")
-
-	blockData := make(map[uint64]blockSummary, 0)
-	for k, b := range blocks {
-		bs := blockSummary{}
-		bs.Block = &blocks[k]
-		bs.Receipts = make(map[ethcommon.Hash]rpctypes.RawTxReceipt, 0)
-		bs.Latencies = make(map[uint64]time.Duration, 0)
-		blockData[b.Number.ToUint64()] = bs
-	}
-
-	for _, r := range txReceipts {
-		bn := r.BlockNumber.ToUint64()
-		bs := blockData[bn]
-		if bs.Receipts == nil {
-			log.Error().Uint64("blocknumber", bn).Msg("Block number from receipts does not exist in block data")
-		}
-		bs.Receipts[r.TransactionHash.ToHash()] = r
-		blockData[bn] = bs
-	}
-
-	nonceTimes := make(map[uint64]time.Time, 0)
-	for _, ltr := range loadTestResults {
-		nonceTimes[ltr.Nonce] = ltr.RequestTime
-	}
-
-	minLatency := time.Millisecond * 100
-	for _, bs := range blockData {
-		for _, tx := range bs.Block.Transactions {
-			// TODO: What happens when the system clock of the load tester isn't in sync with the system clock of the miner?
-			// TODO: the timestamp in the chain only has granularity down to the second. How to deal with this
-			mineTime := time.Unix(bs.Block.Timestamp.ToInt64(), 0)
-			requestTime := nonceTimes[tx.Nonce.ToUint64()]
-			txLatency := mineTime.Sub(requestTime)
-			if txLatency.Hours() > 2 {
-				log.Debug().Float64("txHours", txLatency.Hours()).Uint64("nonce", tx.Nonce.ToUint64()).Uint64("blockNumber", bs.Block.Number.ToUint64()).Time("mineTime", mineTime).Time("requestTime", requestTime).Msg("Encountered transaction with more than 2 hours latency")
-			}
-			bs.Latencies[tx.Nonce.ToUint64()] = txLatency
-			if txLatency < minLatency {
-				minLatency = txLatency
-			}
-		}
-	}
-	// TODO this might be a hack, but not sure what's a better way to deal with time discrepancies
-	if minLatency < time.Millisecond*100 {
-		log.Trace().Str("minLatency", minLatency.String()).Msg("Minimum latency is below expected threshold")
-		shiftSize := ((time.Millisecond * 100) - minLatency) + time.Millisecond + 100
-		for _, bs := range blockData {
-			for _, tx := range bs.Block.Transactions {
-				bs.Latencies[tx.Nonce.ToUint64()] += shiftSize
-			}
-		}
-	}
-
-	printBlockSummary(c, blockData, startNonce, endNonce)
-
-	log.Trace().Str("summaryTime", (endReceipt.Sub(startReceipt)).String()).Msg("Total Summary Time")
-	return nil
-
-}
-
-func isEmptyJSONResponse(r *json.RawMessage) bool {
-	rawJson := []byte(*r)
-	return len(rawJson) == 0
-}
-
-type Latency struct {
-	Min    float64
-	Median float64
-	Max    float64
-}
-
-type Summary struct {
-	BlockNumber uint64
-	Time        time.Time
-	GasLimit    uint64
-	GasUsed     uint64
-	NumTx       int
-	Utilization float64
-	Latencies   Latency
-}
-
-type SummaryOutput struct {
-	Summaries          []Summary
-	SuccessfulTx       int64
-	TotalTx            int64
-	TotalMiningTime    time.Duration
-	TotalGasUsed       uint64
-	TransactionsPerSec float64
-	GasPerSecond       float64
-	Latencies          Latency
-}
-
-func printBlockSummary(c *ethclient.Client, bs map[uint64]blockSummary, startNonce, endNonce uint64) {
-	filterBlockSummary(bs, startNonce, endNonce)
-	mapKeys := getSortedMapKeys(bs)
-	if len(mapKeys) == 0 {
-		return
-	}
-
-	var totalTransactions uint64 = 0
-	var totalGasUsed uint64 = 0
-	p := message.NewPrinter(language.English)
-
-	allLatencies := make([]time.Duration, 0)
-	summaryOutputMode := *inputLoadTestParams.SummaryOutputMode
-	jsonSummaryList := []Summary{}
-	for _, v := range mapKeys {
-		summary := bs[v]
-		gasUsed := getTotalGasUsed(summary.Receipts)
-		blockLatencies := getMapValues(summary.Latencies)
-		minLatency, medianLatency, maxLatency := getMinMedianMax(blockLatencies)
-		allLatencies = append(allLatencies, blockLatencies...)
-		blockUtilization := float64(gasUsed) / summary.Block.GasLimit.ToFloat64()
-		if gasUsed == 0 {
-			blockUtilization = 0
-		}
-		// if we're at trace, debug, or info level we'll output the block level metrics
-		if zerolog.GlobalLevel() <= zerolog.InfoLevel {
-			if summaryOutputMode == "text" {
-				_, _ = p.Printf("Block number: %v\tTime: %s\tGas Limit: %v\tGas Used: %v\tNum Tx: %v\tUtilization %v\tLatencies: %v\t%v\t%v\n",
-					number.Decimal(summary.Block.Number.ToUint64()),
-					time.Unix(summary.Block.Timestamp.ToInt64(), 0),
-					number.Decimal(summary.Block.GasLimit.ToUint64()),
-					number.Decimal(gasUsed),
-					number.Decimal(len(summary.Block.Transactions)),
-					number.Percent(blockUtilization),
-					number.Decimal(minLatency.Seconds()),
-					number.Decimal(medianLatency.Seconds()),
-					number.Decimal(maxLatency.Seconds()))
-			} else if summaryOutputMode == "json" {
-				jsonSummary := Summary{}
-				jsonSummary.BlockNumber = summary.Block.Number.ToUint64()
-				jsonSummary.Time = time.Unix(summary.Block.Timestamp.ToInt64(), 0)
-				jsonSummary.GasLimit = summary.Block.GasLimit.ToUint64()
-				jsonSummary.GasUsed = gasUsed
-				jsonSummary.NumTx = len(summary.Block.Transactions)
-				jsonSummary.Utilization = blockUtilization
-				latencies := Latency{}
-				latencies.Min = minLatency.Seconds()
-				latencies.Median = medianLatency.Seconds()
-				latencies.Max = maxLatency.Seconds()
-				jsonSummary.Latencies = latencies
-				jsonSummaryList = append(jsonSummaryList, jsonSummary)
-			} else {
-				log.Error().Str("mode", summaryOutputMode).Msg("Invalid mode for summary output")
-			}
-		}
-		totalTransactions += uint64(len(summary.Block.Transactions))
-		totalGasUsed += gasUsed
-	}
-	parentOfFirstBlock, _ := c.BlockByNumber(context.Background(), big.NewInt(bs[mapKeys[0]].Block.Number.ToInt64()-1))
-	lastBlock := bs[mapKeys[len(mapKeys)-1]].Block
-	totalMiningTime := time.Duration(lastBlock.Timestamp.ToUint64()-parentOfFirstBlock.Time()) * time.Second
-	tps := float64(totalTransactions) / totalMiningTime.Seconds()
-	gaspersec := float64(totalGasUsed) / totalMiningTime.Seconds()
-	minLatency, medianLatency, maxLatency := getMinMedianMax(allLatencies)
-	successfulTx, totalTx := getSuccessfulTransactionCount(bs)
-
-	if summaryOutputMode == "text" {
-		p.Printf("Successful Tx: %v\tTotal Tx: %v\n", number.Decimal(successfulTx), number.Decimal(totalTx))
-		p.Printf("Total Mining Time: %s\n", totalMiningTime)
-		p.Printf("Total Transactions: %v\n", number.Decimal(totalTransactions))
-		p.Printf("Total Gas Used: %v\n", number.Decimal(totalGasUsed))
-		p.Printf("Transactions per sec: %v\n", number.Decimal(tps))
-		p.Printf("Gas Per Second: %v\n", number.Decimal(gaspersec))
-		p.Printf("Latencies - Min: %v\tMedian: %v\tMax: %v\n", number.Decimal(minLatency.Seconds()), number.Decimal(medianLatency.Seconds()), number.Decimal(maxLatency.Seconds()))
-		// TODO: Add some kind of indication of block time variance
-	} else if summaryOutputMode == "json" {
-		summaryOutput := SummaryOutput{}
-		summaryOutput.Summaries = jsonSummaryList
-		summaryOutput.SuccessfulTx = successfulTx
-		summaryOutput.TotalTx = totalTx
-		summaryOutput.TotalMiningTime = totalMiningTime
-		summaryOutput.TotalGasUsed = totalGasUsed
-		summaryOutput.TransactionsPerSec = tps
-		summaryOutput.GasPerSecond = gaspersec
-
-		latencies := Latency{}
-		latencies.Min = minLatency.Seconds()
-		latencies.Median = medianLatency.Seconds()
-		latencies.Max = maxLatency.Seconds()
-		summaryOutput.Latencies = latencies
-
-		val, _ := json.MarshalIndent(summaryOutput, "", "    ")
-		p.Println(string(val))
-	} else {
-		log.Error().Str("mode", summaryOutputMode).Msg("Invalid mode for summary output")
-	}
-}
-
-func getSuccessfulTransactionCount(bs map[uint64]blockSummary) (successful, total int64) {
-	for _, block := range bs {
-		total += int64(len(block.Receipts))
-		for _, receipt := range block.Receipts {
-			successful += receipt.Status.ToInt64()
-		}
-	}
-	return
-}
-
-func getTotalGasUsed(receipts map[ethcommon.Hash]rpctypes.RawTxReceipt) uint64 {
-	var totalGasUsed uint64 = 0
-	for _, receipt := range receipts {
-		totalGasUsed += receipt.GasUsed.ToUint64()
-	}
-	return totalGasUsed
-}
-
-func getMapValues[K constraints.Ordered, V any](m map[K]V) []V {
-	newSlice := make([]V, 0)
-	for _, val := range m {
-		newSlice = append(newSlice, val)
-	}
-	return newSlice
-}
-
-func getMinMedianMax[V constraints.Float | constraints.Integer](values []V) (V, V, V) {
-	if len(values) == 0 {
-		return 0, 0, 0
-	}
-	sort.Slice(values, func(i, j int) bool {
-		return values[i] < values[j]
-	})
-	half := len(values) / 2
-	median := values[half]
-	if len(values)%2 == 0 {
-		median = (median + values[half-1]) / V(2)
-	}
-	var min V
-	var max V
-	for k, v := range values {
-		if k == 0 {
-			min = v
-			max = v
-			continue
-		}
-		if v < min {
-			min = v
-		}
-		if v > max {
-			max = v
-		}
-	}
-	return min, median, max
-}
-
-func filterBlockSummary(blockSummaries map[uint64]blockSummary, startNonce, endNonce uint64) {
-	validTx := make(map[ethcommon.Hash]struct{}, 0)
-	var minBlock uint64 = math.MaxUint64
-	var maxBlock uint64 = 0
-	for _, bs := range blockSummaries {
-		for _, tx := range bs.Block.Transactions {
-			if tx.Nonce.ToUint64() >= startNonce && tx.Nonce.ToUint64() <= endNonce {
-				validTx[tx.Hash.ToHash()] = struct{}{}
-				if tx.BlockNumber.ToUint64() < minBlock {
-					minBlock = tx.BlockNumber.ToUint64()
-				}
-				if tx.BlockNumber.ToUint64() > maxBlock {
-					maxBlock = tx.BlockNumber.ToUint64()
-				}
-			}
-		}
-	}
-	keys := getSortedMapKeys(blockSummaries)
-	for _, k := range keys {
-		if k < minBlock {
-			delete(blockSummaries, k)
-		}
-		if k > maxBlock {
-			delete(blockSummaries, k)
-		}
-	}
-
-	for _, bs := range blockSummaries {
-		filteredTransactions := make([]rpctypes.RawTransactionResponse, 0)
-		for txKey, tx := range bs.Block.Transactions {
-			if _, hasKey := validTx[tx.Hash.ToHash()]; hasKey {
-				filteredTransactions = append(filteredTransactions, bs.Block.Transactions[txKey])
-			}
-		}
-		bs.Block.Transactions = filteredTransactions
-		filteredReceipts := make(map[ethcommon.Hash]rpctypes.RawTxReceipt, 0)
-		for receiptKey, receipt := range bs.Receipts {
-			if _, hasKey := validTx[receipt.TransactionHash.ToHash()]; hasKey {
-				filteredReceipts[receipt.TransactionHash.ToHash()] = bs.Receipts[receiptKey]
-			}
-		}
-		bs.Receipts = filteredReceipts
-
-	}
-}
-
-func getSortedMapKeys[V any, K constraints.Ordered](m map[K]V) []K {
-	keys := make([]K, 0)
-	for k := range m {
-		keys = append(keys, k)
-	}
-	sort.Slice(keys, func(i, j int) bool {
-		return keys[i] < keys[j]
-	})
-	return keys
 }
 
 func transactOptsToCallMsg(tops *bind.TransactOpts) ethereum.CallMsg {
